Refactor auction module into separate files for better organization
