+++ src/discordbot/cogs/_auction/__init__.py
@@ -1,0 +1,7 @@
+from nextcord.ext import commands
+
+from .cog import AuctionCogs
+
+
+async def setup(bot: commands.Bot) -> None:
+    bot.add_cog(AuctionCogs(bot))
